+++ src/cpp/impl/CellSim.Cells.Cell.cpp
@@ -1,9 +1,9 @@
 ﻿#include "CellSim.Cells.Cell.hpp"
 #include "CellSim.Cells.CellDivisionResult.hpp"
 #include "CellSim.Cells.CellGrowthResult.hpp"
-#include "CellSim.Cells.CellMetabolicArgs.hpp"
-#include "CellSim.Cells.CellMoleculeEmissionArgs.hpp"
 #include "CellSim.Cells.CellMoleculeSensingArgs.hpp"
+#include "CellSim.Cells.MolecularProcessArgs.hpp"
+#include "CellSim.Cells.MolecularProcessResult.hpp"
 #include "CellSim.Messages.hpp"
 #include "CellSim.Molecular.Molecule.hpp"
 #include "CellSim.Molecular.MoleculeInfo.hpp"
@@ -47,17 +47,6 @@ namespace CellSim::Cells
         return m_behaviorPtr->ShouldDivideThisStep(this);
     }
 
-    bool Cell::AppendMolecule(Molecular::MoleculeKind kind)
-    {
-        for (Molecular::Molecule& molecule : m_internalMolecules) {
-            if (molecule.Kind() == kind) [[unlikely]] return false;
-        }
-
-        AppendMoleculeUnsafe(kind);
-
-        return true;
-    }
-
     void Cell::Combine(Cell& c) noexcept
     {
         m_radius = ::cbrt(m_radius * m_radius * m_radius + c.m_radius * c.m_radius * c.m_radius);
@@ -83,25 +72,13 @@ namespace CellSim::Cells
             result.NewDaughter.NewPosition
         );
 
-        for (Molecular::Molecule& molecule : m_internalMolecules) {
-            newDaughter.AppendMoleculeUnsafe(molecule.Kind());
+        // 分子の量を半分にしてコピー
+        for (auto& pair : m_internalMolecules) {
+            pair.second /= 2;
         }
+        newDaughter.m_internalMolecules = m_internalMolecules;
 
         return newDaughter;
-    }
-
-    void Cell::EmitMolecule(Molecular::MoleculeField& field)
-    {
-        Numerics::GridPosition3 position3 = field.ToGridPosition3(m_position);
-
-        field.Concentrations().At(position3.X, position3.Y, position3.Z) += m_behaviorPtr->ComputeMoleculeEmitAmount(this, { &field, position3 });
-    }
-
-    void Cell::EmitMolecule(::std::vector<Molecular::MoleculeField>& fields)
-    {
-        for (Molecular::MoleculeField& field : fields) {
-            EmitMolecule(field);
-        }
     }
 
     void Cell::Grow()
@@ -112,16 +89,31 @@ namespace CellSim::Cells
         m_mass = result.NewMass;
     }
 
-    void Cell::Metabolize()
-    {
-        for (Molecular::Molecule& molecule : m_internalMolecules) {
-            molecule.Amount(molecule.Amount() + m_behaviorPtr->ComputeMetabolicChange(this, { { molecule.Amount(), molecule.Kind() } }));
-        }
-    }
-
     void Cell::Move() noexcept
     {
         m_position += m_force * (Settings::Config::Simulation::DeltaTime() / m_mass);
+    }
+
+    void Cell::ProcessMolecules(Molecular::MoleculeField& field)
+    {
+        Numerics::GridPosition3 position3 = field.ToGridPosition3(m_position);
+
+        Molecular::MoleculeKind kind = field.Kind();
+
+        auto& intracellularAmount = m_internalMolecules[kind];
+        auto& extracellularAmount = field.Concentrations().At(position3.X, position3.Y, position3.Z);
+
+        MolecularProcessResult result = m_behaviorPtr->ComputeMolecularProcess(this, { extracellularAmount, &field, intracellularAmount, kind, position3 });
+
+        intracellularAmount += result.IntracellularChange;
+        extracellularAmount += result.ExtracellularChange;
+    }
+
+    void Cell::ProcessMolecules(::std::vector<Molecular::MoleculeField>& fields)
+    {
+        for (Molecular::MoleculeField& field : fields) {
+            ProcessMolecules(field);
+        }
     }
 
     void Cell::SenseMolecules(Molecular::MoleculeField const& field)