+++ src/cpp/impl/CellSim.Cells.NormalCellBehavior.hpp
@@ -7,16 +7,23 @@
 namespace CellSim::Cells
 {
     class NormalCellBehavior : public CellBehavior {
+        private:
+        /// @brief 細胞の半径がこの値を超えると分裂する
+        double m_cellDivisionRadius;
+
         public:
+
+        NormalCellBehavior();
+
+        NormalCellBehavior(double cellDivisionRadius);
 
         CellDivisionResult ComputeDivisionOutcome(const Cell* sender) override;
         CellGrowthResult ComputeGrowth(const Cell* sender) override;
-        double ComputeMetabolicChange(const Cell* sender, CellMetabolicArgs args) override;
-        double ComputeMoleculeEmitAmount(const Cell* sender, CellMoleculeEmissionArgs args) override;
+        MolecularProcessResult ComputeMolecularProcess(const Cell* sender, MolecularProcessArgs args) override;
         
         CellBehavior* CreateClone() const override;
         
-        constexpr bool HasState() const noexcept override;
+        constexpr bool IsReusable() const noexcept override;
         Numerics::Vector3 OnSenseMolecules(const Cell* sender, CellMoleculeSensingArgs args) override;
         bool ShouldDivideThisStep(const Cell* sender) noexcept override;
     };
@@ -26,12 +33,12 @@ namespace CellSim::Cells
 {
     inline CellBehavior* NormalCellBehavior::CreateClone() const
     {
-        return new NormalCellBehavior();
+        return new NormalCellBehavior(*this);
     }
 
-    constexpr bool NormalCellBehavior::HasState() const noexcept
+    constexpr bool NormalCellBehavior::IsReusable() const noexcept
     {
-        return false;
+        return true;
     }
 }